front-server/requirements.txt
aio-pika==6.8.0
aiohttp==3.7.3
aiormq==3.3.1
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
gunicorn==20.0.4
idna==3.1
idna-ssl==1.1.0
multidict==5.1.0
pamqp==2.3.0
PyYAML==5.4.1
typing-extensions==3.7.4.3
yarl==1.6.3
